+++ edflow/hooks/logging_hooks/tensorboard_handler.py
@@ -1,5 +1,4 @@
 import numpy as np
-import tensorboardX
 import yaml
 from edflow.iterators.batches import batch_to_canvas
 
@@ -17,6 +16,13 @@ def log_tensorboard_images(writer, results, step, path):
         v = batch_to_canvas(v)
         v = ((v + 1) * 127.5).astype(np.uint8)
         writer.add_image(tag=k, img_tensor=v, global_step=step, dataformats="HWC")
+    writer.flush()
+
+
+def log_tensorboard_figures(writer, results, step, path):
+    results = dict((path + "/" + k, v) for k, v in results.items())
+    for k, v in results.items():
+        writer.add_figure(tag=k, figure=v, global_step=step)
     writer.flush()